requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.10.0
elementpath==2.0.5
pkg-resources==0.0.0
python-magic==0.4.18
pytz==2020.4
sqlparse==0.4.1
xmlschema==1.3.1
pytest-django==4.1.0